app_backend/myapp/requirements/base.txt
# Django and its libraries
Django==3.1.4
django-extensions==3.1.0
django-cors-headers==3.6.0
django-countries==7.0
django-model-utils==4.1.1
django-filter==2.4.0
social-auth-app-django==4.0.0
django-fernet-fields==0.6

# DB
psycopg2==2.8.6

# Monitoring
django-health-check==3.16.1
psutil==5.8.0

# GraphQL
graphene-django==2.15.0
graphene-django-tools==0.21.0
django-graphql-jwt==0.3.1
django-graphql-social-auth[jwt]==0.1.4

# WSGI
gunicorn==20.0.4

# ASGI
channels==3.0.3

# Storages
boto3==1.16.47
django-storages==1.11.1

# Linting
flake8==3.8.4

# Testing
coverage==5.3.1
factory_boy==3.2.0
pytest==6.2.1
pytest-cov==2.10.1
pytest-django==4.1.0
parameterized==0.7.4
responses==0.12.1
django-debug-toolbar==3.2

# Misc
pytz==2020.5
requests==2.25.1
Pillow==8.0.1
money==1.3.0
django-sendgrid-v5==0.9.0

# Background task
django-redis==4.12.1
